+++ .dagger/build.go
@@ -239,25 +239,6 @@ func goModule() *dagger.Go {
 		WithBuildCache(cacheVolume("go-build"))
 }
 
-func goModuleCross(platform dagger.Platform) *dagger.Go {
-	container := goModule().
-		WithCgoEnabled(). // TODO: set env var instead?
-		Container().
-		With(func(c *dagger.Container) *dagger.Container {
-			if platform != "" {
-				c = c.WithEnvVariable("TARGETPLATFORM", string(platform))
-			}
-
-			return c
-		}).
-		WithDirectory("/", dag.Container().From(xxBaseImage).Rootfs()).
-		WithExec([]string{"apk", "add", "--update", "--no-cache", "ca-certificates", "make", "git", "curl", "clang", "lld"}).
-		WithExec([]string{"xx-apk", "add", "--update", "--no-cache", "musl-dev", "gcc"}).
-		WithExec([]string{"xx-go", "--wrap"})
-
-	return dag.Go(dagger.GoOpts{Container: container})
-}
-
 func (m *Build) HelmChart(
 	// Name of the chart to build.
 	name string,
