+++ .github/workflows/sdk-javascript-release.yaml
@@ -26,7 +26,7 @@ jobs:
         id: get-short-sha
         run: |
           id=$(echo ${{github.sha}} | cut -c 1-12)
-          echo "::set-output name=id::$id"
+          echo "id=${id}" >> $GITHUB_OUTPUT
 
       - name: Publish NPM package
         uses: dagger/dagger-for-github@e47aba410ef9bb9ed81a4d2a97df31061e5e842e # v8.0.0
