+++ .vscode/settings.json
@@ -5,7 +5,7 @@
   "go.testEnvVars": {
     "TZ": "UTC",
     "POSTGRES_HOST": "127.0.0.1",
-    "OPENMETER_ADDRESS": "http://127.0.0.1:8888",
+    "OPENMETER_ADDRESS": "http://127.0.0.1:38888",
     // Useful when the schema is evolving during development (currently only supported
     // by billing)
     // "TEST_DISABLE_ATLAS": "true"
