+++ e2e/.gitignore
@@ -1,0 +1,2 @@
+logs
+docker-compose.override.yaml
