+++ e2e/docker-compose.openmeter-latest.yaml
@@ -1,0 +1,8 @@
+services:
+  openmeter:
+    image: ghcr.io/openmeterio/openmeter:latest
+    pull_policy: always
+
+  sink-worker:
+    image: ghcr.io/openmeterio/openmeter:latest
+    pull_policy: always
