+++ quickstart/.gitignore
@@ -1,0 +1,1 @@
+docker-compose.override.yaml
